refactor: unify theme setting across activities
