+++ app/src/main/java/com/cappielloantonio/tempo/ui/activity/LoginActivity.kt
@@ -1,7 +1,6 @@
 package com.cappielloantonio.tempo.ui.activity
 
 import android.content.Intent
-import android.content.res.Configuration
 import android.os.Bundle
 import android.widget.ArrayAdapter
 import android.widget.Toast
@@ -16,11 +15,9 @@ import androidx.media3.common.util.UnstableApi
 import com.cappielloantonio.tempo.App
 import com.cappielloantonio.tempo.R
 import com.cappielloantonio.tempo.databinding.ActivityLoginBinding
-import com.cappielloantonio.tempo.helper.ThemeHelper
 import com.cappielloantonio.tempo.model.Server
-import com.cappielloantonio.tempo.util.Preferences
+import com.cappielloantonio.tempo.util.ActivityUtil
 import com.cappielloantonio.tempo.viewmodel.ServerViewModel
-import com.google.android.material.color.DynamicColors
 
 
 class LoginActivity : AppCompatActivity() {
@@ -32,7 +29,7 @@ class LoginActivity : AppCompatActivity() {
     private var isInitialSyncDone = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        enableTheme()
+        ActivityUtil.enableThemeSwitch(this)
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
 
@@ -51,42 +48,12 @@ class LoginActivity : AppCompatActivity() {
         setupOldLoginButton()
     }
 
-    private fun enableTheme() {
-        val theme: String = Preferences.getTheme()
-        val darkStyle: String = Preferences.getDarkThemeStyle()
-        val isAmoled = ThemeHelper.AMOLED_MODE == darkStyle
-        var applyAmoled = false
-
-        if (ThemeHelper.DARK_MODE == theme || ThemeHelper.AMOLED_MODE == theme) {
-            if (isAmoled) {
-                setTheme(R.style.AppTheme_Amoled)
-                applyAmoled = true
-            }
-        } else if (ThemeHelper.DEFAULT_MODE == theme) {
-            val nightModeFlags =
-                getResources().configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
-            if (nightModeFlags == Configuration.UI_MODE_NIGHT_YES && isAmoled) {
-                setTheme(R.style.AppTheme_Amoled)
-                applyAmoled = true
-            }
-        }
-
-        DynamicColors.applyToActivityIfAvailable(this)
-        if (applyAmoled) {
-            getTheme().applyStyle(R.style.ThemeOverlay_App_Amoled, true)
-        }
-    }
-
     private fun initEdgeToEdge() {
         ViewCompat.setOnApplyWindowInsetsListener(binding.main) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
-    }
-
-    private fun initTheme() {
-
     }
 
     @OptIn(UnstableApi::class)
